See release notes
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@97041 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: aa834b14d73b8a8f7df8a0175751f6e0151d2007

+++ Online/Online/GaudiUPI/src/PropertyMenu.cpp
@@ -99,7 +99,7 @@ Gaudi::PropertyMenu::PropertyMenu(const std::string& nam, IInterface* iface, Int
     }
     int cmd = CMD_SHOW+1;
     if ( vlen > 32 ) vlen = 32;
-    sprintf(fmt,"%%-%lds%%-%lds%%-%lds %%-3s %%-3s",nlen+2,tlen+2,vlen+2);
+    sprintf(fmt,"%%-%lds%%-%lds%%-%lds %%-3s %%-3s",long(nlen+2),long(tlen+2),long(vlen+2));
     m_window->addCOM(cmd,fmt,"Property Name","Type","Value","CBR","CBU");
     for(std::vector<Property*>::const_iterator i=prps.begin(); i != prps.end(); ++i)  {
       const Property* p = (*i);
